requirements.txt
google-generativeai
crewai
crewai[tools]
feedparser
atproto
requests
python-dotenv
streamlit
openai